Difference between sums of odd and even digits
Given a number S in the form of string, the task is to check whether the sum of digits at odd places is equal to the sum of digits at even places or not.

Example 1:

Input: S = "132"
Output: 1
Explanation: The sum of digits at odd 
places is 1+2=3. Similarly the sum of 
digits at even places is 3. Since they 
are equal,the answer is 1.