Confused pappu
Pappu is confused between 6 & 9. He works in the billing department of abc company and his work is to return the remaining amount to the customers. If the actual remaining amount is given we need to find the maximum possible extra amount given by the pappu to the customers.

Example 1:

Input: amount = 56
Output: 3
Explanation: maximum possible extra 
             amount = 59 - 56 = 3.