Tough Competitors
There are n competitors in an exam. Each competitor has his own skill value which is given by the array arr where arr1 is the skill of the first competitor, arr2 is the skill of second competitor and so on. Two competitors are said to be tough competitors if their skill difference is least i.e. they are very close in their skill values. Given n and an array arr as input, find the tough competitors among the n competitors and print the absolute of the difference of their skill values.

Example 1:

Input:
n = 4
arr[] = {9, 4, 12, 6}
Output: 2