Cricket Average
It’s time to find out which player is consistent in cricket. You need to calculate the average of a player based on his score. Player average is defined as total runs scored by a player in N matches divided by the number of matches in which he gets out.
 

Example 1:

Input: run = {10, 101, 49}, 
status = {"out", "notout", "out"}
Output: 80