Average Count Array
You are given an integer 'n' which denote the number of elements in an array a[ ] and an integer 'x'. You have to calculate the average of element a[i] and x and find out if that number exist in array or not. Do it for all the elements of array and store the count result in another array for each index i.

Note: Always take the floor value of the average.

Example 1:

Input : arr[ ] = {2, 4, 8, 6, 2} and X = 2
Output : 2 0 0 1 2