Print an array in Pendulum Arrangement
Given an array arr of size n. Arrange the elements of the array in a way similar to the to-and-fro movement of a Pendulum.

The minimum element out of the list of integers, must come in center position of array. If there are even elements, then minimum element should be moved to (n-1)/2 index (considering that indexes start from 0)
The next number (next to minimum) in the ascending order, goes to the right, the next to next number goes to the left of minimum number and it continues like a Pendulum.
As higher numbers are reached, one goes to one side in a to-and-fro manner similar to that of a Pendulum
Example 1:
Input : 
n = 5
arr[] = {1, 3, 2, 5, 4}
Output :
5 3 1 2 4