Repeated IDs
Geek is given a task to select at most 10 employees for a company project. Each employee is represented by a single digit ID number which is unique for all the selected employees for the project. Geek has a technical problem in his system which printed ID number multiple times. You are given array a[ ] of N elements having all printed IDs. Help him to get rid of the repeated IDs.

Example 1:

Input:
N = 5
a[ ] = {8, 8, 6, 2, 1}
Output:
8 6 2 1