Ladoo problem
Bheem has N friends. And he has a ladoo for each day. Given an Array A[] denoting on which day ith friend wants a ladoo. If Bheem is unable to give a ladoo to a friend, he loses his friendship with them. Find out the maximum number of Friends he can have at the end.

 

Example 1:

Input:
N = 5
A[] = {3, 3, 1, 2, 4}
Output:
4