Triangle growing downwards
Given string str of a constant length, print a triangle out of it. The triangle should start with the first character of the string and keep growing downwards by adding one character from the string.
The spaces on the left side of the triangle should be replaced with a dot character. 

Example 1:

Input:
str = "geeks"

Output:
....g
...ge
..gee
.geek
geeks