India's unique railway tracks
There are infinite number of cities in Geekland numbered from 1 onward. Some of them are connected via bidirectional railway tracks.
There are railway tracks 1-2, 3-4, 5-6, 7-8, and so on.
There are railway tracks between every two consecutive odd numbers. 1-3, 3-5, 5-7, 7-9, and so on.
There are railway tracks between every two consecutive even numbers. 2-4, 4-6, 6-8, 8-10, and so on.
Given two cities a and b, check if there is a railway track between them or not.


Example 1:

Input: 
a = 1, b = 4
Output: NO