Encoding names
CBI agents are investigating a case in which they came across certain names of the culprits. They decided to encode the names into the number format.

"Riya" is encoded as "729611390","Sumitha" as "73108101981109993" ,"Anandita" as "55101891039410011294" and so on. Given a String S denoting a name, Encode the name.

 

Example 1:

Input:
S = "Soni"
Output:
7310210298