The problem of identical arrays
Two arrays of size N are called identical arrays if they contain the same elements. The order of elements in both arrays could be different; however, both the arrays must contain same elements. You are given two arrays of size N.

You need to determine if the arrays are identical or not.

Example 1:
 

Input : A[] = {1, 2, 3, 4, 5}, B[] = {3, 4, 1, 2, 5}
Output : 1