Same frequency in two halves
Given a string S, find whether it fulfills the following criteria. 
When split from the middle, the string should give two halves having the same characters and same frequency of each character. If the number of characters in the string is odd, ignore the middle character.


Example 1:

Input: S = "abcdbca"
Output: YES